// src/types/geoDB.types.ts
// Generated by https://quicktype.io

export interface GeoDBCityResponse {
  data: Datum[];
  links: Link[];
  metadata: Metadata;
}

export interface Datum {
  id: number;
  wikiDataId: string;
  type: string;
  name: string;
  country: string;
  countryCode: string;
  region: string;
  regionCode: number | string;
  latitude: number;
  longitude: number;
}

export interface Link {
  rel: string;
  href: string;
}

export interface Metadata {
  currentOffset: number;
  totalCount: number;
}